+++ packages/cad-simple-viewer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/cad-simple-viewer
 
+## 1.0.16
+
+### Patch Changes
+
+- feat: add new property 'background' for component MlCadViewer
+
 ## 1.0.15
 
 ### Patch Changes